requirements-test.txt
# Testing dependencies for AI Image Detector
pytest>=7.0.0
pytest-cov>=4.0.0
pytest-mock>=3.10.0
pytest-xdist>=3.0.0  # For parallel test execution

# Core dependencies (should match main requirements)
numpy>=1.21.0
Pillow>=8.0.0

# Optional dependencies for testing
opencv-python>=4.5.0  # For testing cv2 integration
torch>=1.9.0  # For testing torch integration (optional)
transformers>=4.20.0  # For testing transformers integration (optional)
timm>=0.6.0  # For testing timm integration (optional)
